+++ telicent-graph-schema/src/main/java/io/telicent/jena/graphql/fetchers/telicent/graph/AbstractPagingFetcher.java
@@ -66,6 +66,7 @@ public abstract class AbstractPagingFetcher<TSource, TInput, TOutput> implements
         List<Filter> filters = new ArrayList<>();
         if (this.enableFilters()) {
             createFilters(environment, filters);
+            filters.removeIf(f -> f instanceof IncludeAllFilter);
         }
 
         return Txn.calculateRead(dsg, () -> {