+++ telicent-graph-schema/src/main/java/io/telicent/jena/graphql/fetchers/telicent/graph/AbstractRelationshipsFetcher.java
@@ -1,11 +1,11 @@
 /**
  * Copyright (C) Telicent Ltd
- *
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
- *
+ * <p>
  * http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
@@ -21,11 +21,13 @@ import org.apache.jena.atlas.lib.tuple.Tuple4;
 import org.apache.jena.graph.Node;
 import org.apache.jena.sparql.core.DatasetGraph;
 import org.apache.jena.sparql.core.Quad;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.Set;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 /**
@@ -35,6 +37,8 @@ import java.util.stream.Stream;
  */
 public abstract class AbstractRelationshipsFetcher<TOutput>
         extends AbstractPagingFetcher<TelicentGraphNode, Quad, TOutput> {
+    private static final Logger LOGGER = LoggerFactory.getLogger(AbstractRelationshipsFetcher.class);
+
     /**
      * The direction of relationships we're configured to fetch
      */
@@ -148,6 +152,12 @@ public abstract class AbstractRelationshipsFetcher<TOutput>
      * @return Stream of quads representing relationships
      */
     private Stream<Quad> streamPreFiltered(DatasetGraph dsg, TelicentGraphNode node, List<Tuple4<Node>> quadPatterns) {
+        if (LOGGER.isDebugEnabled()) {
+            LOGGER.debug("Translated pre-filter eligible filters into following {} Quad Patterns:\n  {}",
+                         quadPatterns.size(),
+                         quadPatterns.stream().map(Object::toString).collect(Collectors.joining("\n  ")));
+        }
+
         Stream<Quad> stream = quadPatterns.stream().flatMap(p -> dsg.stream(p.get(0), p.get(1), p.get(2), p.get(3)));
         return switch (this.direction) {
             case OUT -> stream.filter(q -> q.getObject().isURI() || q.getObject().isBlank());